refactor(council): give inline findings and the consensus round their own modules
The source-size ratchet refused the previous commit's shape, correctly. Inline
findings charged +153 lines across orchestrator.ts, council_cli.ts and
ai_council/config.ts — all three far past the 1500-line ceiling, where every
added line counts. Paying that by shortening comments would have deleted the
reasoning and kept the bloat, so it is paid by extraction instead, in two steps
because the first was not enough.

ai_council/inline_findings.ts now holds the whole concern: the trailing-block
locator and its span type, the RecordedExtractionOutcome union,
harvest_inline_findings, composeFinalRoundSuffix, takeInlineFindings and
inlineFindingsActive. That is where they belong on merit and not only on
budget — the contract's promise, the locator, the harvest and the config
predicate all have to agree, and a reader should meet them together. It took
the charge to +23; the residual was irreducible call sites.

ai_council/consensus_round.ts then takes run_consensus_scoring — the function
this change actually edits — out of orchestrator.ts. orchestrator.ts
re-exports the name, so every existing importer resolves unchanged. The new
module imports consult back, closing a module cycle that is call-time rather
than init-time (nothing there runs at module load), which is the shape ESM
resolves without a partial-initialisation hazard.

composeFinalRoundSuffix absorbs the stance-contract branch it now composes
with. The order it fixes is load-bearing rather than cosmetic: the stance
contract requires the stance to be the reply's last line, so a findings block
appended after it would contradict the instruction the member is reading.

Net for the tree: -174 lines. The baseline is lowered to the exact live total
measured on the merged tree, not left loose, because
check_source_size_budget.test.ts asserts baseline equals tree total — that
test is how the looseness surfaced, while the gate itself only advised.

Behaviour is unchanged: 1069 council tests green, all five sensitivity arms
re-run against the moved code and each observed red before restore.
